Rework of top level CMake code related to Dakota. Changed BUILD_WITH_DAKOTA to REQUIRE_DAKOTA, and moved REQUIRE_DAKOTA to an advanced option to better reflect how Dakota is a standalone and optional tool.
Now, we always look for Dakota, but only complain about not finding it if REQUIRE_DAKOTA is set to true (as it will be for the continuous integration machines).

Also added version handling. Set the required version of Dakota to 5.3.1 so I can try out the new version with our unit tests.

As a later step, I plan to conditionally compile the analysisdriver unit tests that use Dakota based on whether Dakota is actually present on the system.

[#52020843]