Add ModelObjectList model files
The cpp, hpp, and *_Impl.hpp files for ModelObjectList generated by the
ruby script. The hpp and *_Impl.hpp have a rough first attempt at the
required methods for this model object. Also in the OpenStudio.idd the
OS:ModelObjectList was moved from the' \group OpenStudio HVAC' to
'\group OpenStudio Resources'