[#54312158] Clean up superbuild and tests for Exception handling bugs on MacOS
The crux of the problem is that we are linking OpenStudio/Boost/Qt with SDK 10.7
while Ruby is linked with SDK 10.8. This seems to 'confuse' the runtime system
such that it is unable to recognize derived exception types during exception
handling.

The previous use of the ruby_cpp executable in OpenStudio for executing ruby tests
was masking this problem. Presumably because it was linked to the same version of
the sdk that openstudio was linked to. All investigations point to this.

I've changed openstudiocore/ruby/CMakeLists.txt to no longer use RUBY_CPP and
instead use RUBY_EXECUTABLE on Unix, when executing Ruby unit tests to make
sure we do not mask problems like this in the future.

The exact nature of the problem is not fully understood. However, our options
for moving forward are:

  1. Link to SDK 10.8 and set SDK 10.5 (min version supported by Qt) as the
     min-version. This would mean we only officially support 10.8, but older
     versions may work.
  2. Make two different builds, one for 10.7, one for 10.8
  3. Use our ruby_cpp internally for launching scripts, instead of the system executable.
     This would still leave open the problem of using OpenStudio via the system ruby.
  4. Build and ship our own version of Ruby on MacOS. This has the same problem as #3.

My vote is for #1. Only officially support 10.8 and re-evaluate when 10.9 is released.

@kbenne @rpg123 @nllong @axelstudios the main Mac users and maintainers,
please weigh in.