Update SDD translator to support the SZVAV tag.
When AirSys elements have the SZVAVAC or SZVAVHP type, they will be
translated with a SetpointManagerWarmest and a VAV terminal regardless
of the specified air terminal type.

Generally the SDD will define an uncontrolled terminal for SZVAV, but we have
determined that the most appropriate modeling strategy is to use a
reheat terminal.