Create new process to monitor cloud
A new thread has been created to watch the cloud and make sure the UI
reflects the correct state.  Later the monitor process can do other
things, like issue dialogs if the internet drops out, or the server is
still running but the openstudio-server application drops out.

There are still some loose ends to cleanup, like timeout for example,
but this is a stable framework to build on.