[#54854900] Get JSON unit tests for flatoutdir passing
@elainethale this should get everything working for updating of remote
jobs. It definitely requires a build on the server/worker nodes. Once I
got into it, I realized it really is a problem specific to flatoutdir
being enabled, so I made the solution specific to that.

The change set is pretty minor. If it's a flatoutdir, I save the
directory name that was used, the re-use that directory name based
against the basePath when returning outdir() from the job.

I updated this branch to be up to date with develop, not sure what, if
anything, you can review before a build is made for the worker nodes.