Create string based IddObjectType
Replaced OpenStudio Enum, IddObjectType, with a new class that is based
on a string.  This removes the need to coordinate IDD objects types in a
central location, laying the groundwork to do incremental building when
new IDD objects are added.  Only building openstudio_utilies and related
dependencies.  And there might be bugs fallout from this.

One thing that is lost is int IddObjectType::value().  If someone can
devise a hash function based on std::string IddObjectType::valueName()
we could get ::value() back.  ::value() is commonly used in switch
statements on IddObjectType.

Also IddObjectType::Foo where foo is an object name is deprecated.
These have to be centrally located since you can't reopen a c++ class
and that central coordination breaks incremental building which is the
whole point of this.  Existing classes have IddObjectType::Foo method,
but new entries in the Idd will not create new members in IddObjectType.
Use IddObjectType(iddobjectname::Foo) instead.