SDD: Generate "offset" schedules for min vent
When translating OASchMthd where the method is FollowAvailability, we
clone the availability schedule and apply start and stop offsets.  In
this way the ventilation follows the availability schedule of the system
with a start and stop lag, useful during morning warmup for instance.

One limitation of this implementation is that it is built around
ScheduleYear -> ScheduleWeek -> ScheduleDay.  If the ScheduleWeek
references the same ScheduleDay for several days of the week, this clone
method is not sophisticated enough to recognize the shared ScheduleDay.
The cloned ScheduleWeek will have unique schedules for each day of the
week.