SDD: Adjust zone sizing parameters
This removes default values for zone sizing heating and cooling air flow
rates.  Only values from the SDD xml file will be used.  The design air
flow method will be set to DesignDay unless the SDD specifies a heating
or cooling air flow, in which case the method is DesignDayWithLimit.