[59295112] Get OpenStudio working with Ruby mingw32 build
@macumber

Notes:

  - The build of ruby we're using is a slightly modified version of the ruby-installer.org 2.0.0-p247. It's been enhanced to include a linker def file for MSVC and it's config.h file has been replaced with one from an MSVC build of ruby, so that the #defines are correct for MSVC
  - The build is currently hosted on my dropbox account. We'll need to work with @axelstudios once we are comfortable with it to get it put on NREL's server
  - I have verified that OpenStudio, Pat and SketchUp can all launch with these ruby bindings. OpenStudio and Pat successfully use the embedded ruby interpreter
  - SketchUp fails to launch the plugin because of .rb script changes that need to be made.

Final Notes

I cannot explain the great trouble I had with the official mingw build last time I tried. My only though is that it was because I was working with a 64 bit build at the time.

With this current success it likely makes sense to drop our aws specific build of ruby and unify on this build of it at some point. We would need to add the required modules and then replace the config.h as descript in the README in the ruby.zip file I posted.