Final cleanup of new algorithm
The new schedule translation algorithm passes all the tests that
we currently have. It uses the daysFromFirstReport vector, which
simplifies some things but doesn't solve the precision issue. The
same algorithm should also work for the variable interval
schedule, but it would probably be a good idea to rework the
TimeSeries internals and then revisit this and the variable
interval case.

[#60441426]