Update and fix Fan:OnOff
FanOnOff had various bugs in its implementation.

When the FanOnOff is removed, previously it would not remove() the
curves associated with it due to missing children() function.

FanOnOff used inheirited addToNode() implementation, which allowed it
to be added to an AirLoopHVAC. This is not possible in EnergyPlus.
FanOnOff can only be added to a containing HVAC Component, such as
Unitary, ZoneHVAC, etc.

The constructor required two curves to be passed into it but then
created two curves within the constructor, never using the passed in
curves. This has been changed.

FanOnOff within EnergyPlus does not require the two curves, which allow
the change from constant speed to multi speed. However, the current
OpenStudio implemenation requires these curves thus a new overloaded
constructor is made which creates constant curves and does not require
the curves to be passed into the constructor. This allows FanOnOff to
be modeled as constant speed.

The unit tests for FanOnOff are also updated to have better test
coverage.