Update AirChiller, remove RefrigerationChillerSet
Updates AirChiller to be a ZoneHVACComponent and removes
ZoneHVAC:RefrigerationChillerSet. In E+, the air chillers are added to
the chiller set which is added to the equipment list. In OS, we will
avoid the chiller set and make air chillers able to be added to
equipment list then in forward translator create chiller set and add
air chillers to chiller set.