SDD: Improve four pipe fan coil translation
This is really an improvement of the components within the fan coil.
For FanOnOff translate fan power curve.  For heating and cooling water
coils, translate the design inlet / outlet temperatures.

In the case of water coils, this change will impact all coils, not just
those on a ZnSys.  Previously design inlet / outlet temperatures where
completely defaulted, now they are based on appropriate properties in
the containing ZnSys or AirSys.