Update Refrigeration:AirChiller
Changes AirChiller from using node connections to not using them. This
is changed because using node connections left orphaned nodes in E+ IDF
due to the way AIrChiller is handled in Forward Translation. This is
more in line with current E+ (8.0) usage of AirChiller where
ZoneHVAC:RefrigerationChillerSet does not currently use nodes either.