Remove extraneous airLoopHVAC() calls and casts
In forward translate sizing zone, it was making a call to airLoopHVAC()
which cycles through all components and takes a while if there are many
components on the loop. Plus the variable isn’t even being used.

In HVACComponent, airLoopHVAC() and plantLoop() both loop through a
vector of their respective classes but then optionalCast to the same
class. It is not necessary.