Fix miscellaneous errors related to plenums
This is a collection of bug fixes that result in functioning plenum
support with clean simulations.  Included in these bug fixes are the
following:

In ForwardTranslateAirLoopHVACReturnPlenum and
ForwardTranslateAirLoopHVACSupplyPlenum, fix issue where wrong node name
is used for inlet/outlet nodes.

In AirLoopHVACSupplyPlenum fix issue in overloaded addBranchForZone that
can cause infinite recursion.

In all AirTerminal classes, relax logic in ::addToNode to use base
Splitter / Mixer type instead of the more specific
AirLoopHVACZoneSplitter/Mixer.  This allows ::addToNode to function
properly between plenums.  This error was not captured by unit tests,
but resulted in simulation errors.  Similar addToNode logic is present
in all air terminal classes and this would be a good area to factor a
base class and remove this code duplication in the future.

In ModelHVAC.i, reorder class registration to properly capture base type
of plenums.