Attempt to catch zombied processes
I've been able to reproduce situations where EnergyPlus would exit
because of a signal but the QProcess object would not know that the
process exited.

This seems to occur only if the parent process is catching signals for
the children, like when being executed from within Node. It's possible
that other situations could cause similar problems, such as Ruby or R
calling OpenStudio.