Avoid creatName for Connection objects
The Connection ModelObject works behind the scenes to join HVAC
components together.  A typical model has a very large number of them
and they are all getting a pretty human readable name for no purpose.
This eats up time.  Consider setting name to the handle string and avoid
work of the createName method.

Note that Connection has overloaded createName for some time but it is
not working.  It seems virtual functions are not called in constructors.

The method used here is to give the name field a default and then set
name to the handle string in the Connection constructor.  The default
implementation of createName seems to avoid the expensive operation if
there is a default.

[#66666852]