Update Case and Walkin forward translators
This implements DefrostCycleParameters simple schedule generation in
each of the forward translators for Case and Walkin.

There is a new Impl only function in Case and Walkin,
defrostStartTimes(). This function returns an ordered vector of Times
if they are set.

This also eliminates some bugs in createSimpleSchedule and adds
documentation to createSimpleSchedule header file.