Fix bug in createSimpleSchedule
Schedule:Compact requires the schedule to have a value for all 24
hours. Also, Schedule:Compact requires the last value to be 24:00,
which differs from Openstudio::Time which represents it as 00:00.

So now if defrost1StartTime doesn’t equal 00:00 or is less than 24,
Case and Walkin create an Openstudio::Time for 24:00.

createSimpleSchedule now translates 00:00 to 24:00 for use in
Schedule:Compact and makes sure it is the last value.