Remove write to HVACAutoSizing property
This property was written to the SDD as 1, meaning autosize.  Now it is
not written to the SDD file at all so it is up to CBECC to put it in.
Keep in mind that the SDD->OS translator still defaults to autosize if
there is no HVACAutoSizing property in the SDD.