Fix ProposedEnergy+ idd error
AirLoopHVAC:UnitarySystem has four fields which duplicate the field
names of four different fields within the object. This causes an issue
with the enumeration setup in OS. According to the IO Reference, these
four fields should have ‘When No Cooling or Heating is Required’ at the
end of the field names.