Add DX coil translateAndMapModelObject 'magic'
This change allows translateAndMapModelObject to use the correct
forward translator for DX coils depending on whether or not they are
part of a containing system, such as a unitary system.

This checks if the coil is on an AirLoopHVAC (and thus use
translateFoo) or if the coil is in a containing component (and thus use
translateFooWithoutUnitary).