Remove hardcoded stdlib on OSX
This removes the hardcoded -stdlib=libstdc++ on OSX which was needed
for boost 1.47.

The new default stdlib on OSX 10.9 is libc++, however, libc++ is
included on OSX back to 10.7.

With this change, Openstudio will use the default stdlib depending on
the OSX target.

This change is also needed if Openstudio will use any c++11 features on
OSX since it requires the use of libc++.