Bugfixes and updates
This removes the ability of the user to set and use the ControlType.
ControlType can be either Load or SetPoint, if it is SetPoint you need
to define SetpointManagers for each node within the Unitary System
which a user can no do from the GUI. So we are commenting out this
functionality for now.

This also fixes a small bug in forward translation and changes the mini
icon to an actual mini icon.