Improve AirLoopHVAC::zoneMixer / zoneSplitter
If there are components between demand inlet or demand outlet nodes and
zone mixer or splitter, the
implementation of zoneMixer and zoneSplitter would not work.  Normally
there are not components in these locations, but it might happen in some
rare or future configurations.  In this case it was actually a bug where
cooling coils were allowed on the demand side that exposed this problem
with zoneSplitter / Mixer methods.  The method should be robust enough
to handle this configuration even if it is rarely if ever encountered.
For one the model has been (incorrectly) allowing coils on the demand
side, but then this issue with the splitter and mixer methods caused a
crashed when trying to remove the coils, making the minor bug a major
bug.

[#68356360]