Update StraightComponent::addToNode
StraightComponent::addToNode has gradually become more capable, allowing
these types of components to be added in more locations.  For example
StraightCompononent can now add coponents to the demand side of
AirLoopHVAC.  This is good for the generic base class implementation,
The problem is that the derived components have not beenproperly keeping
pace to dissallow addToNode in appropriate context.  So for instance
CoilHeatingDXSingleSpeed unintentionally gained the ability to show up
on the AirLoopHVAC demand side.

This work corrects on error with StraightCompononent::addToNode, when
the demand side is empty.  Basically fully committing to the generic
power of StraightCompononent::addToNode.  Additionally derived components
were updated to restrict addToNode as required, before blindly calling
base class's generic algorithm.

[#68356360]