Fix clone bug in some components
The bug in clone…
If a component on a loop is cloned then added to a loop, addToNode
would fail and return false. This shouldn’t happen since this is a
freshly cloned object.

The issue stemmed from reimplemented clone methods in some objects
derived from StraightComponent using ModelObject::clone instead of
StraightComponent::clone. The main different between these two clones
is that StraightComponent resets the inletPort and outletPort on clone.
This was not happening with ModelObject::clone, hence the addToNode
issue with the cloned object.

This also updates all the unit tests to now test for this case.