Fix openstudio_utilities_tests link issue
The CMake macro CREATE_TEST_TARGETS had the target link libraries
ordered with gtest and gtest_main before the thread library, and
on Ubuntu 14.04 this was causing the link to fail for
openstudio_utilities_tests. Moving ALL_DEPENDENCIES to after gtest
and gtest_main fixes that. Also, the thread library was showing
up twice in the link command, so I commented out the explicit
addition for the utilities test. I've left it in as a comment
since I can only build this branch on Ubuntu and it needs to be
verified on other platforms.