Patch swig's ruby generator to avoid memory error
The static data in the ruby modules appears to conflict when more than
one ruby module is loaded on MacOS. The exact cause is undetermined, but
the code in question did not enter SWIG until 3.0, and the problem only
seems to exist on Xcode 5.1.

Debugging indicates that the objects in question were in fact created
in different memory locations even before the anonymous namespace
was added. Furthermore, debugging indicated no changes in the memory
layout were affected by this change.

However, the change fixes the issue and shows no averse effects.

Apple's runtime linking of static objects has always been confusing
to me, and this is no different.

Regardless, the change should be harmless as the anonymous namespace
ensures that the objects are local to an individual translation unit
and I believe that's the intention here. Inter-module data sharing in
SWIG should occur through mechanisms provided by the scripting
language's runtime itself, not through static data.