Add Deprecated.hpp for compile time warning
This allows for compile time warning of deprecated functions. It is
used as follows…

OS_DEPRECATED functionName();

You also should add…
LOG(Warn, “ThisClass::thisFunction has been deprecated and will be
removed in a future release, please use OtherClass::otherFunction”);

to the implementation part of the deprecated function.

The ‘\deprecated’ tag should be added to the doxygen comments for the
deprecated function.

This also suppresses the deprecated warnings for unit tests and swig
bindings to reduce compile noise.