Limit new setpoint control option
Access to AirLoopHVACUnitarySystem setpoint control is now limited to
private impl only.  This is because it only works with single components
in the unitary.  Inner components still do not receive setpoint
managers, and that is not expected to change on this branch.  It might
be something better to handle in the event that inner components have
proper nodes and connections, sometime in the future.  This would be the
so called "thickwalled" approach to components with inner components.

[#73904900]