Fix translation bug in HeatExchangerFluidToFluid
The availability schedule of HVAC components is no longer required by
EnergyPlus.  If it is not specified, EnergyPlus will default to always
on.  The OpenStudio code for HeatExchangerFluidToFluid had confusion in
this area.  The model object was initialized with an always on schedule,
while the method returned optional Schedule.  No problem there, except
the E+ translator does not check the optional before tranlsating,
resulting in a crash.

Removed the code to initialize availability schedule in
HeatExchangerFluidToFluid.  Also updated the translator to check if
there is an availability schedule before attempting to translate the
schedule.  This better reflects current E+ convention and new objects
should follow this patter.

[#76734866]