Add support for staged systems
In particular this adds support for:

CoilCoolingDXMultiSpeed
CoilHeatingGasMultiStage
ZoneControlThermostatStagedDualSetpoint
AirLoopHVACUnitaryHeatPumpAirToAirMultiSpeed

There is a new base class, Thermostat, which
ZoneControlThermostatStagedDualSetpoint and
ThermostatSetpointDualSetpoint derive from.

There is no method to author new instances of this system through the
UI, only API.  However, once staged systems are created they can be
manipulated in the user interface.  Need to add library entries for UI,
and also develop an interface to interact with staged thermostats.  This
work on thermostats will overlap with development on ThermalZone grid
view.

[#75620048]