Enable queuing of refresh events.
This prevents refreshes from occuring more often then they need to. It's
still not the granularity that we need, but refreshes are generally
collapsed to one when many changes happen at the same time.

Something appears to be wrong with the main tab of the thermalzones
grid - like an object is not currently hooked up. Will investigate this
next.

[#74774412] @evanweaver