Improve logic to generate mixed air SPMs
With the introduction of humidity based setpoint managers, the logic to
generate mixed air setpoint managers was compromised.  A humidity based
setpoint manager on a particular node would block the generation of
temperature SPMs even though it is still required.  This new logic
checks to see if an SPM is required by checking not only for an SPM, but
a temperature based SPM.