Get deletion of drop zone items working
@evanweaver The failure case was because of the second mechanism in
which an OSItem may be constructed, if there wasn't already one assigned
to the m_item member.

Please review this to see if it makes sense. I don't known when one
mechanism is used over another.

Also, the reset function was not being hooked up in GridController. It
now is.