Fix crash when inspecting zone from grid view
Best bet is to not allow inspecting the thing that is in the grid, since
that is the point of the grid.  But in case you do, for instance inspect
a thermal zone from the thermal zone grid view, then avoid showing
plenum interface (outside the hvac tab) because plenum colors come from
the hvac view.

[#80334374]