Change default "Heat Recovery Bypass Control Type"
In ControllerOutdoorAir change the default value for
field "Heat Recovery Bypass Control Type" from
"BypassWhenOAFlowGreaterThanMinimum" to
"BypassWhenWithinEconomizerLimits".

The old value was majorly messing up oa rates even when there is no heat
recovery device.  This corrects unmet hours in CBECC lab model, which is
approximatly system 7, but with exahaust fans and 100% oa.  Many other
models could be inflicted with this problem, which ultimately is an
EnergyPlus bug documented here:
https://github.com/NREL/EnergyPlus/issues/4568.  The scope of this issue
in OpenStudio models is unclear but is being investigated with the
EnergyPlus issue 4568.  It may or may not be isolated to 100% OA and/or
exhaust fans.

[#81583558]