Adjust PIU fan schedule
When setting AirLoopHVAC::availabilitySchedule or when adding a new piu
terminal to an existing AirLoopHVAC, make sure the piu fan matches the
system availability schedule.  This is especially useful when night
cycle option is used on the AirLoopHVAC system.

A user can still set the fan availabilty schedule directly to ovrride
this default behavior.

[#84632652]