Add unit tests around SpaceType::clone
All of these tests appear to show expected behavior, however the issue
described in #870 remains.  The fact is the behavior described in the
issue is as designed even if it is not very desirable in this case.
Consider a feature or change in behavior to address #870.

Calling ResourceObject::clone from Model A to Model B will always result
in the creation of a new object even if that object already exists in
the target Model, and even if it is a ResourceObject.  In contrast, when
a ModelObject that *references* ResourceObject instance(s) is cloned,
duplicates of the ResourceObject instance(s) will not be made.

This is all demonstrated more clearly than these words in the unit tests added
here.

[#69514744]