Add tests for exhaust and inlet port lists
ThermalZone objects are attached to inlet "port lists" and exhaust "port
lists" because they can have multiple connections at those locations.
The ThermalZone class has ThermalZone::exhaustPortList and
ThermalZone::inletPortList that both return the PortList class.
PortList keeps a reference back to the associated ThermalZone and
ThermalZone keeps a reference back to the ThermalZone so that the lookup
is fast in either direction.  User model for issue #447 had a missing
reference from PortList back to ThermalZone.  It is unknown why this
situation occured (model was from version 1.0) but here unit tests are
added to verify that the issue no longer exists.