Change StraightComponent::addToNode
StraightComponent::addToNode will now remove the component from any
existing loop and then add to the new node.  Previously the method would
simply return false, which is inconsistent with other HVACComponent
instances such as WaterToAircomponent.  This brings consistency to the
API while also taking the opportunity to greatly consolidate the connect
/ disconnect logic into one place in a private HVACComponent method.

With this work comes a new method StraightComponent::removeFromLoop
which mirrors similar methods in other HVACComponent types such as
WaterToAirComponent::removeFromAirLoop and ::removeFromPLantLoop.
Previously it was only possible to completely remove a StraightComponent
from the model.  ::removeFromLoop disconnects it from the system but
leaves it in the model.

close #1658