Add support for tertiary plant loops
Some WaterToWaterComponent types can interact with a third system.  One
example of this is the ChillerAbsorptionIndirect.  New methods have been
added to support this including ::addToTertiaryNode,
::tertiaryPlantLoop, ::removeFromTertiaryPlantLoop.

The low level search algorithms such as Loop::supplyComponents() have
been updated to account for tertiary systems, and this has altered the
implementation and signature of the underlying private methods HVACComponent::edges
and Loop::findModelObject(s).