Log instead of assert for translate HVAC
When translating HVAC systems or compound objects favor logging instead
of asserting when a referenced component is missing.  For example if a
PTAC references a fan that is not found in the SDD or fails to translate
log the event and fail to create the PTAC instead of asserting on the
missing fan.