Fix simulation failure for CAV/VAV, no fan sch
Fixes #1605

This solves the issue with CAV and VAV fans with a blank Supply Air Fan
Operating Model Schedule. These hvac_library objects will simulate now.

It is currently defaulted to an Always On Discrete schedule, but the
user can change a schedule in the GUI where they can not change a Fan
Type.