Remove AvailabilityManagerScheduled and AvailabilityManagerAssignmentList
These are no longer used.  We support one availability manager at a
time.  The availability schedule is now attached directly to the system, and
an AvailabilityManagerScheduled is created on idf creation (based on
availability schedule) if there is not another more advanced
availability manager applied.  This is not a change in functionality
(actually there are more AVM options now) we are just removing the
intermediate connecting objects that clutter things up for features that
(laryering AVM) that we are not going to use.