Reorganize translate plant op schemes
This is the beginning of major rework for plant operation scheme
translation. Recently explicit ModelObject types were added for the
various operation schemes and translators have already been put in
place. Soon those translators will be utilitized (called from) this new
ForwardTranslatePlantEquipmentOperationSchemes function. The defaulting
logic is the focus here in this commit. That is even though there are
explicit model object types for the operation schemes we still want the
old defaulting logic to work. This commit is an effort to retain that
default logic but clean it up in a way that is more maintainable. Most
of this content is pulled out of TranslatePlantLoop, but reworked in
signifcant ways. This will not be the final word on this effort. More
commits to come.