Finish reorganize translate plant op schemes
Translate plant op schemes now happen in an isolated function
ForwardTranslatePlantEquipmentOperationSchemes instead of within the
translatePlantLoop function. This is nicer and includes a bunch of other
improvements to maintainability. One side effect of this is that op
schemes now come at the end of the plant input in the idf file. It
formerly came up near the top right after PlantLoop. New position is not
as ideal but motivated by some coding structure conveniences. Consider
moving back up later.

As described in previous commit, this change supports translating
explicit ModelObject op schemes, but retains the defaulting mechanism if
nothing is defined in the Model. Assume there will be additional commits
to clean up loose ends.